refactor: move cowswap hooks out of common-hooks and remove unused hooks
