+++ apps/cowswap-frontend/src/cosmos.decorator.tsx
@@ -1,12 +1,11 @@
 import '@reach/dialog/styles.css'
 import './polyfills'
 
-import React, { StrictMode, useCallback, useContext, ReactNode, useEffect } from 'react'
+import React, { ReactNode, StrictMode, useCallback, useContext, useEffect } from 'react'
 
 import IMAGE_MOON from '@cowprotocol/assets/cow-swap/moon.svg'
 import IMAGE_SUN from '@cowprotocol/assets/cow-swap/sun.svg'
-import { BlockNumberProvider } from '@cowprotocol/common-hooks'
-import { WalletUpdater, injectedWalletConnection } from '@cowprotocol/wallet'
+import { injectedWalletConnection, WalletUpdater } from '@cowprotocol/wallet'
 import { Web3ReactProvider } from '@web3-react/core'
 
 import { LanguageProvider } from 'i18n'
@@ -14,12 +13,13 @@ import SVG from 'react-inlinesvg'
 import { Provider } from 'react-redux'
 import { HashRouter } from 'react-router-dom'
 import { Flex } from 'rebass'
-import styled from 'styled-components/macro'
-import { ThemeContext } from 'styled-components/macro'
+import styled, { ThemeContext } from 'styled-components/macro'
 
 import { cowSwapStore } from 'legacy/state'
 import { useDarkModeManager } from 'legacy/state/user/hooks'
 import ThemeProvider, { FixedGlobalStyle, ThemedGlobalStyle } from 'legacy/theme'
+
+import { BlockNumberProvider } from './common/hooks/useBlockNumber'
 
 const DarkModeToggleButton = styled.button`
   display: flex;
