+++ apps/cowswap-frontend/src/legacy/components/Header/NetworkSelector/index.tsx
@@ -151,7 +151,6 @@ const NetworkAlertLabel = styled.div`
   }
 `
 
-
 export function NetworkSelector() {
   const provider = useWalletProvider()
   const { chainId } = useWalletInfo()
